+++ src/fdc_ts/core.clj
@@ -51,19 +51,23 @@
 
 ;TODO Move put to separate module
 
-(defn- auth [token ctx]
-  (= (token env) (get-in ctx [:request :headers "auth-token"])))
+(defn- auth [token project-tokens project ctx]
+  (def request-token  (get-in ctx [:request :headers "auth-token"]))
+  (or (= (token env) request-token)
+      (= (get-in (json/parse-string (project-tokens env))  [project]) request-token)))
 
 (defn- auth-configured [token ctx]
   (token env))
 
-(def auth-publish (partial auth :auth-token-publish))
-(def auth-publish-configured (partial auth-configured :auth-token-publish))
+(def auth-publish (partial auth :auth-token-publish :auth-token-project))
+(def auth-publish-configured (and(partial auth-configured :auth-token-publish)
+                                 (partial auth-configured :auth-token-project)))
 
-(def auth-statistics (partial auth :auth-token-statistics))
-(def auth-statistics-configured (partial auth-configured :auth-token-statistics))
+(def auth-statistics (partial auth :auth-token-statistics :auth-token-project))
+(def auth-statistics-configured (and (partial auth-configured :auth-token-statistics)
+                                     (partial auth-configured :auth-token-project)))
 
-(def auth-meta (partial auth :auth-token-meta))
+(def auth-meta (partial auth :auth-token-meta "{}" nil ))
 (def auth-meta-configured (partial auth-configured :auth-token-meta))
 
 (def project-malformed? (comp not validate-project-data :json))
@@ -76,7 +80,7 @@
   :available-media-types ["application/json"]
   :allowed-methods [:put]
   :service-available? auth-publish-configured
-  :authorized? auth-publish
+  :authorized? (partial auth-publish nil)
   :allowed? (comp main-project-exists? :json)
   :put! (fn [ctx] 
           (when (not (project-exists? (:json ctx))) (add-project (:json ctx)))
@@ -86,14 +90,14 @@
   :available-media-types ["application/json"]
   :allowed-methods [:get]
   :service-available? auth-statistics-configured
-  :authorized? auth-statistics
+  :authorized? (partial auth-statistics project)
   :handle-ok (fn [_] (handle-coverage-request time project subproject language)))
 
 (defresource get-project-coverage-diff [project days]
   :available-media-types ["application/json"]
   :allowed-methods [:get]
   :service-available? auth-statistics-configured
-  :authorized? auth-statistics
+  :authorized? (partial auth-statistics project)
   :handle-ok (fn [_] (json/generate-string (project-diff-days project days))))
 
 (defresource put-project []